move messages to one component
